+++ docs/ChangeLog.md
@@ -3,6 +3,8 @@
 ## v 1.0.0 - initial version released
 ## v 1.0.1 - Issue #1 Cache lookup table values after initial load (big performance improvement)
 ## v 1.0.2 - Issue #2 Optionally output updated script via stdout (and not updating source/destination files) by using -StandardOutput parameter
-## v 1.0.3 - Issue #7 Support PowerShell Core
+## v 1.0.3 - Issue #7 Initial support for PowerShell Core
 ## v 1.0.4 - Issue #13 Add space after . when dot-sourcing a command
 ## v 1.0.5 - Issue #7 and #16 - missing aliases across OSes
+## v 1.0.6 - Issue #7 Core - only support known safe cross-OS aliases
+             Tested on Windows PowerShell, (Windows OS) PowerShell Core, Ubuntu 16.04 (Docker image)